moving chunking file into socket_transport
